requirements.txt
apache-airflow-providers-amazon==8.7.0
apache-airflow-providers-postgres==5.7.0
boto3==1.34.90
mlflow==2.15.1
pandas==2.0.3
psycopg2-binary==2.9.6
scikit-learn==1.3.0
torch==2.0.1
transformers==4.30.1
tensorboard==2.13.0
openpyxl==3.1.2
